Introduced new attribute - state
The state attribute, in its current form takes
one of two values - active or dormant. A project
is said to be active if the last commit made to
the project is within a specified threshold.

The duration since last commit is computed
relative to the current date i.e. datetime.today()
or a configured value allowing for the
specification of the date the GHTorrent dump was
downloaded. MAX(commits.created_at) could have
been used but the current MAX(commits.created_at)
has 2038 as its year.

Change resolves #8