Using MAX(commits.created_at) instead of .SHA
The SHA of the most recent commit from `commits`
proved to be unreliable. `git reset --hard SHA`
would intermittently fail with "Basd object"
error. Resorting to using the created_at of the
most recent commit with `git log --before` to
get the SHA of the commit closest to the one in
the database.